CMA-62 [refactor] : JWT 관련 createAccessToken 에서 role인자를 배열에서 상수 enum 으로 변경
