+++ src/main/resources/application.yaml
@@ -7,7 +7,7 @@ spring:
 # spring.datasource.url=jdbc:mysql://todayCheckMysql:3306/todaycheck?useSSL=false&allowPublicKeyRetrieval=true&characterEncoding=UTF-8&serverTimezone=UTC
     driver-class-name: com.mysql.cj.jdbc.Driver
     username:
-    password: 
+    password:
   h2:
     console:
       enabled: true
