+++ src/test/java/com/team/comma/repository/UserRepositoryTest.java
@@ -2,6 +2,8 @@ package com.team.comma.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.team.comma.constant.UserRole;
+import com.team.comma.constant.UserType;
 import java.util.Collections;
 
 import org.junit.jupiter.api.DisplayName;
@@ -10,8 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
-import com.team.comma.entity.User;
-import com.team.comma.entity.User.UserType;
+import com.team.comma.domain.User;
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE) // EmbeddedDatabase 가 아닌 Mysql 에 테스트 사용됨
@@ -53,8 +54,8 @@ public class UserRepositoryTest {
 	}
 
 	private User getUserEntity() {
-		return User.builder().email(userEmail).password(userPassword).userType(UserType.GeneralUser)
-				.roles(Collections.singletonList("ROLE_USER")).build();
+		return User.builder().email(userEmail).password(userPassword).type(UserType.GeneralUser)
+				.role(UserRole.USER).build();
 	}
 
 }
